requirements.txt
# Core dependencies - required for all installations
boto3>=1.26.0
jmespath>=1.0.0  # Required by boto3
pandas>=1.5.0
pyarrow>=10.0.0
pyyaml>=6.0
requests>=2.28.0
tenacity>=8.0.0
pydantic>=2.0.0
ibis-framework[duckdb]>=9.0.0  # Portable dataframe library with DuckDB backend
fsspec>=2023.1.0  # Filesystem abstraction for S3/local storage
s3fs>=2023.1.0  # S3 filesystem implementation for fsspec
pyodbc>=4.0.0  # ODBC support for database extractors

# Note: Optional dependencies moved to separate files:
# - requirements-azure.txt: Azure Blob Storage support
# - requirements-db.txt: Database extractor support (SQL Server, PostgreSQL, MySQL)
# - requirements-async.txt: Async HTTP extractor support
#
# Install with: pip install -r requirements.txt -r requirements-azure.txt
# Or use extras: pip install -e .[azure,db,async]
